# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, e.g. geo fields missing for license geo, or sent for another license"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an event (you become its owner; an invite code is generated)",
        "tags": [
          "Events (Track Vote)"
        ]
      },
      "get": {
        "operationId": "EventsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventViewDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Public events + the events you are a member of, newest first",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
    "/events/join": {
      "post": {
        "operationId": "EventsController_joinByCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No event with this code"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join any event with its 8-character invite code (as guest; never downgrades your role)",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_view",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Event details, your role, and whether you can vote (send ?lat=&lng= for geo events)",
        "tags": [
          "Events (Track Vote)"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner can do this"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an event (owner). Switching license to geo requires all geo fields",
        "tags": [
          "Events (Track Vote)"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner can do this"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an event and its queue (owner)",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
    "/events/{id}/join": {
      "post": {
        "operationId": "EventsController_joinPublic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a public event as guest",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
    "/events/{id}/invites": {
      "post": {
        "operationId": "EventsController_invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner can do this"
          },
          "404": {
            "description": "Event not found, or no account with this email"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a registered account by email (owner). Invited accounts can vote under license invited",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Music Room API",
    "description": "REST API of Music Room (Track Vote events, accounts, music search).\n\n**Auth:** log in with `POST /auth/login` (or a social login), then click **Authorize** and paste the `accessToken`.\nAccess tokens last 15 minutes; get a new pair with `POST /auth/refresh`.\n\n**Real-time** (Socket.IO namespace `/events`): see `docs/realtime.md` in the repository.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The accessToken from /auth/login"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "license": {
            "type": "string",
            "enum": [
              "open",
              "invited",
              "geo"
            ]
          },
          "geoLat": {
            "type": "number"
          },
          "geoLng": {
            "type": "number"
          },
          "geoRadiusM": {
            "type": "number",
            "minimum": 10,
            "maximum": 50000
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "visibility"
        ]
      },
      "EventGeoDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "lat": {
            "type": "number",
            "example": 48.8966
          },
          "lng": {
            "type": "number",
            "example": 2.3185
          },
          "radiusM": {
            "type": "number",
            "example": 200
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "lat",
          "lng",
          "radiusM"
        ]
      },
      "UserRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ]
      },
      "ParticipationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "description": "Present when `allowed` is false.",
            "enum": [
              "not_member",
              "not_invited",
              "not_started",
              "ended",
              "location_required",
              "outside_area"
            ],
            "type": "string"
          },
          "allowed": {
            "type": "boolean",
            "description": "Whether you can vote and suggest tracks right now."
          }
        },
        "required": [
          "allowed"
        ]
      },
      "EventViewDto": {
        "type": "object",
        "properties": {
          "visibility": {
            "enum": [
              "public",
              "private"
            ],
            "type": "string"
          },
          "license": {
            "description": "open = anyone who can see it · invited = owner + invited accounts · geo = inside an area during a time window",
            "enum": [
              "open",
              "invited",
              "geo"
            ],
            "type": "string"
          },
          "geo": {
            "nullable": true,
            "description": "Only when license is 'geo'.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventGeoDto"
              }
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/UserRefDto"
          },
          "myRole": {
            "nullable": true,
            "description": "Your membership; null = not a member (only possible for public events).",
            "enum": [
              "owner",
              "invited",
              "guest"
            ],
            "type": "string"
          },
          "inviteCode": {
            "type": "string",
            "nullable": true,
            "description": "8 characters; only shown to members.",
            "example": "KVAMDDE4"
          },
          "participation": {
            "description": "For license 'geo' this depends on the location you sent (?lat=&lng=).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipationDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Friday Party"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "visibility",
          "license",
          "geo",
          "owner",
          "myRole",
          "inviteCode",
          "participation",
          "createdAt",
          "id",
          "name",
          "description"
        ]
      },
      "JoinByCodeDto": {
        "type": "object",
        "properties": {
          "inviteCode": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8
          }
        },
        "required": [
          "inviteCode"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "license": {
            "type": "string",
            "enum": [
              "open",
              "invited",
              "geo"
            ]
          },
          "geoLat": {
            "type": "number"
          },
          "geoLng": {
            "type": "number"
          },
          "geoRadiusM": {
            "type": "number",
            "minimum": 10,
            "maximum": 50000
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          }
        }
      },
      "InviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "InviteResultDto": {
        "type": "object",
        "properties": {
          "invited": {
            "$ref": "#/components/schemas/UserRefDto"
          }
        },
        "required": [
          "invited"
        ]
      },
...
    }
  }
}